add missing @dur
one note in the Example_Authority_data.mei was missing the dur attribute
an no dur.default is being defined in the file.
With this commit an assumed value for dur is added to the note.
Unfortunately there was no access to the "original", so the value is
based upon the time signature an the dur values of the other notes of
the corresponding layer.