Make Doc_starts_with_music.mei valid
The doument is an extract of Header_nonFRBR.mei but in order to be valid against the associated schema v2.1.1/mei-CMN.rng control events may not be nested inside staff elements.
This commit fixes the issue by moving the control events into the ancestor measure-element and adding a corresponding staff-attribute on each control event.